fix(gpu): 32-bit flat indexing made the col-major PCA kernels fail past 2.1e9 elements
`scale_columns_kernel` and `mean_correct_colmajor_strided_kernel` computed
`int total = m * k` and indexed with a 32-bit `idx`, while the host launched a
grid sized in `usize`. `n_obs × k` crosses 2^31 at census scale — 36M cells ×
k=60 = 2.16e9 — and `randomized_pca_core`'s VRAM pre-flight admits that shape
on an 80 GB H100, so this was reachable, not theoretical.

`int total = m * k` is signed overflow, i.e. UB, so the manifestation is the
compiler's choice. Measured against the pre-fix kernels at 2^31 + 256 elements
(nvcc 12.6, -O3, compute_70, H100): CUDA_ERROR_ILLEGAL_ADDRESS — threads whose
`idx` wrapped negative survive the bounds check and write below the buffer,
which poisons the CUDA context for the process. Past 2^32 elements the host's
`(total as u32)` block count is the silent variant instead: a grid too small to
reach the tail, no error at all. Either way the work does not happen, on the
two paths that matter — `zero_center=True` yielding an uncentered PCA
(gpu_pca_resident.rs passes shard_rows = n_obs), and the final embedding
leaving unscaled by its singular values.

Widen the three col-major kernels to `long long` parameters with 64-bit totals
and flat indices, matching spmm_mean_correct.cu, the row-major sibling that
always had this right. `outer_sub_kernel` gets the same treatment; n_vars × k
is ~3e6 today, so that one is defence-in-depth.

Delete `select_top_eigvecs_desc_kernel` and `reverse_tail_vec_kernel`: dead
since the GPU covariance-PCA path was removed, and the first carried the
identical overflow. A dead broken kernel is worth less than the diff to fix it.

Host side, add `scx_gpu::flat_launch_1d(total: u64, threads: u32)` — the
extraction of the one already-correct copy of this arithmetic, in
`mean_correct_gpu` — and route five truncating `(total as u32).div_ceil(...)`
sites through it. It counts in u64 and returns `ShapeMismatch` when the block
count exceeds the driver's `grid_dim.x` cap rather than casting a wrapped value.

Also promote the two warp-per-row ids the review names as latent siblings
(sparse_dense.cu, normalize_log1p.cu::normalize_warp_row) and
`row_scale_kernel`'s global `row_offset + row`. Those launch 32 threads per row,
so their flat index wraps at 134.2M rows; per-shard row counts are far below
that, so this is defence-in-depth and the comments say so.

Addresses review §8.4.

+++ scx-gpu/kernels/normalize_log1p.cu
@@ -106,7 +106,12 @@ extern "C" __global__ void row_scale_kernel(
     long long start = indptr[row];
     long long end = indptr[row + 1];
 
-    float factor = factors[row_offset + row];
+    // 64-bit global row id (review §8.4): `factors` is indexed in the *global*
+    // obs row space, so `row_offset + row` is an n_obs-scale quantity even
+    // though each shard's `n_rows` is small. The host already refuses a
+    // row_offset past i32::MAX (`gpu_preprocess.rs`); this makes the kernel
+    // correct on its own terms rather than by the caller's grace.
+    float factor = factors[(long long)row_offset + row];
     for (long long i = start; i < end; i++) {
         data[i] *= factor;
     }
@@ -189,8 +194,13 @@ __device__ __forceinline__ void normalize_warp_row(
     bool do_log1p
 ) {
     int lane = threadIdx.x % warpSize;
-    int warp_global = (blockIdx.x * blockDim.x + threadIdx.x) / warpSize;
-    if (warp_global >= n_rows) return;
+    // 64-bit warp id (review §8.4): one warp per row means the flat thread index
+    // is 32 × n_rows, which wraps `unsigned` past 134.2M rows; the negative id
+    // then passes the `>= n_rows` guard and `indptr[warp_global]` reads out of
+    // bounds. Per-shard row counts are far below that, so this is
+    // defence-in-depth — the nonzero addressing below is already 64-bit.
+    long long warp_global = ((long long)blockIdx.x * blockDim.x + threadIdx.x) / warpSize;
+    if (warp_global >= (long long)n_rows) return;
 
     long long start = indptr[warp_global];
     long long end = indptr[warp_global + 1];
